refactor: split env into dedicated module
the @inlang/shared module is a fellacy and will be unbundled with a follow up PR
