+++ source-code/env-variables/src/privateEnv.ts
