+++ source-code/env-variables/src/validateEnvVariables.ts
