+++ source-code/env-variables/tsconfig.json
@@ -1,10 +1,11 @@
 {
 	"extends": "../../tsconfig.base.json",
-	"include": ["lib/**/*.ts"],
+	"include": ["src/**/*.ts", "src/**/*.ts"],
 	"compilerOptions": {
 		"lib": ["ESNext", "DOM"],
 		"types": [],
 		"outDir": "./dist",
-		"rootDir": "./lib"
+		// this module is not bundling. thus, set to esnext
+		"target": "ESNext"
 	}
 }
