+++ source-code/env-variables/vitest.config.ts
@@ -1,5 +1,5 @@
-import { privateEnv } from "./lib/env/index.js"
-import { definePublicEnvVariables } from "./lib/env/src/definePublicEnvVariables.js"
+import { privateEnv } from "./src/index.js"
+import { definePublicEnvVariables } from "./src/definePublicEnvVariables.js"
 import { defineConfig } from "vitest/config"
 
 // no config needed
