+++ source-code/shared/lib/rpc/index.ts
@@ -1,5 +1,4 @@
 import { rpcClient } from "typed-rpc"
-import { isDevelopment } from "../env/index.js"
 
 // ! Only import the type to not leak the implementation to the client
 import type { RpcService } from "./src/index.js"
@@ -13,5 +12,5 @@ const route = "/shared/rpc"
  * This is used by the client to call RPC functions.
  */
 export const rpc = rpcClient<RpcService>(
-	(isDevelopment ? "http://localhost:3000" : "https://inlang.com") + route,
+	(process.env.DEV ? "http://localhost:3000" : "https://inlang.com") + route,
 )
