--- source-code/shared/src/README.md
@@ -1,1 +1,0 @@
-Shared code for the libraries in [lib](../lib)
