+++ source-code/shared/tsconfig.build.json
@@ -1,0 +1,8 @@
+// Exclude test and source map files when building for production
+{
+	"extends": "./tsconfig.json",
+	"exclude": ["**/*.test.ts", "**/*.test-d.ts"],
+	"compilerOptions": {
+		"sourceMap": false
+	}
+}
